refactor: apply review feedback (background layer logic and docs)
